+++ examples/resources/cockroach_private_endpoint_trusted_owner/cockroach_private_endpoint_trusted_owner.tf
@@ -1,0 +1,9 @@
+variable "cluster_id" {
+  type = string
+}
+
+resource "cockroach_private_endpoint_trusted_owner" "example" {
+  cluster_id        = var.cluster_id
+  type              = "AWS_ACCOUNT_ID"
+  external_owner_id = "012345678901"
+}