+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- New cockroach_private_endpoint_connection resource allows users to configure
+  trusted owner entries for private endpoints.
+
 ## [1.1.0] - 2023-08-15
 
 ### Added